Move the filtering tests to a new FacultyPage.test.tsx file
Since the filtering logic was moved up to the FacultyPage component, the logic no longer fits in the FacultyTable suite of tests.
