+++ src/client/components/pages/Faculty/__tests__/FacultyPage.test.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import {
   appliedMathFacultyScheduleResponse,
+  metadata,
   notActiveACSFacultyScheduleResponse,
   partiallyActiveAMFacultyScheduleResponse,
 } from 'testData';
@@ -11,7 +12,9 @@ import {
   QueryByText,
   fireEvent,
   within,
+  wait,
   GetByText,
+  AllByRole,
 } from 'test-utils';
 import { deepStrictEqual, strictEqual } from 'assert';
 import {
@@ -22,7 +25,8 @@ import {
 import * as dummy from 'testData';
 import { FacultyAPI } from 'client/api';
 import { MetadataContextValue } from 'client/context';
-import { TERM } from 'common/constants';
+import { ABSENCE_TYPE, FACULTY_TYPE, TERM } from 'common/constants';
+import { absenceEnumToTitleCase, facultyTypeEnumToTitleCase } from 'common/utils/facultyHelperFunctions';
 import FacultySchedule from '../FacultyPage';
 
 describe('Faculty Page', function () {
@@ -133,6 +137,332 @@ describe('Faculty Page', function () {
         }
       );
       strictEqual(getStub.args[1][0], nextAcademicYear);
+    });
+  });
+  describe('Filtering data', function () {
+    let metadataContext: MetadataContextValue;
+    let getByLabelText: BoundFunction<GetByText>;
+    let getAllByRole: BoundFunction<AllByRole>;
+    let queryAllByRole: BoundFunction<AllByRole>;
+    const testFacultySchedules = [
+      dummy.appliedMathFacultyScheduleResponse,
+      dummy.electricalEngineeringFacultyScheduleResponse,
+      dummy.computerScienceFacultyScheduleResponse,
+    ];
+    const areaFilterLabel = 'Change to filter the faculty list by area';
+    const lastNameLabel = 'Change to filter the faculty list by last name';
+    const firstNameLabel = 'Change to filter the faculty list by first name';
+    const categoryLabel = 'Change to filter the faculty list by category';
+    const fallAbsenceLabel = 'Change to filter the faculty list by the fall absence value';
+    const springAbsenceLabel = 'Change to filter the faculty list by the spring absence value';
+    beforeEach(async function () {
+      getStub.resolves(testFacultySchedules);
+      metadataContext = new MetadataContextValue({
+        ...dummy.metadata,
+        currentAcademicYear,
+        semesters: [
+          `${TERM.FALL} 2020`,
+          `${TERM.SPRING} 2021`,
+          `${TERM.FALL} 2021`,
+          `${TERM.SPRING} 2022`,
+        ],
+      },
+      spy());
+      ({ getByLabelText, getAllByRole, queryAllByRole } = render(
+        <FacultySchedule />, { metadataContext }
+      ));
+      await wait(() => getAllByRole('row').length > 3);
+    });
+    describe('Area Filter', function () {
+      context('when no filters have been changed', function () {
+        it('should default to the value "All"', function () {
+          const rows = getAllByRole('row');
+          const utils = within(rows[2]);
+          const filter = utils
+            .queryByLabelText(areaFilterLabel) as HTMLSelectElement;
+          strictEqual(filter.value, 'All');
+        });
+        it('should show all of the areas and "All" as options', function () {
+          const rows = getAllByRole('row');
+          const utils = within(rows[2]);
+          const filter = utils.queryByLabelText('Change to filter the faculty list by area') as HTMLSelectElement;
+          const options = within(filter)
+            .getAllByRole('option') as HTMLOptionElement[];
+          const actualOptions = options
+            .map(({ textContent }) => textContent);
+          const expectedOptions = metadata.areas;
+          // The default value of the area dropdown is 'All'
+          expectedOptions.unshift('All');
+          deepStrictEqual(actualOptions, expectedOptions);
+        });
+        it('should show all the rows', function () {
+          const tableBodyRows = queryAllByRole('row')
+            .filter((row) => (
+              within(row).queryAllByRole('columnheader').length === 0
+            ));
+          strictEqual(tableBodyRows.length, testFacultySchedules.length);
+        });
+      });
+      context('when the filter is changed', function () {
+        it('should show the rows containing the area selected', function () {
+          const testArea = testFacultySchedules[0].area;
+          const filter = getByLabelText(areaFilterLabel);
+          fireEvent.change(
+            filter,
+            {
+              target: {
+                value: testArea,
+              },
+            }
+          );
+          const expectedNumRows = testFacultySchedules
+            .filter((faculty) => faculty.area === testArea).length;
+          const tableBodyRows = queryAllByRole('row')
+            .filter((row) => (
+              within(row).queryAllByRole('columnheader').length === 0
+            ));
+          const actualNumRows = tableBodyRows.map((row) => (
+            row.textContent
+          )).length;
+          strictEqual(expectedNumRows, actualNumRows);
+        });
+      });
+    });
+    describe('Last Name Filter', function () {
+      context('when no filters have been changed', function () {
+        it('should show all the rows', function () {
+          const tableBodyRows = queryAllByRole('row')
+            .filter((row) => (
+              within(row).queryAllByRole('columnheader').length === 0
+            ));
+          strictEqual(tableBodyRows.length, testFacultySchedules.length);
+        });
+      });
+      context('when the filter is changed', function () {
+        it('should show the rows with last names matching the filters', function () {
+          const testLastName = testFacultySchedules[0].lastName;
+          const filter = getByLabelText(lastNameLabel);
+          fireEvent.change(
+            filter,
+            {
+              target: {
+                value: testLastName,
+              },
+            }
+          );
+          const expectedNumRows = testFacultySchedules
+            .filter((faculty) => faculty.lastName === testLastName).length;
+          const tableBodyRows = queryAllByRole('row')
+            .filter((row) => (
+              within(row).queryAllByRole('columnheader').length === 0
+            ));
+          const actualNumRows = tableBodyRows.map((row) => (
+            row.textContent
+          )).length;
+          strictEqual(expectedNumRows, actualNumRows);
+        });
+      });
+    });
+    describe('First Name Filter', function () {
+      context('when no filters have been changed', function () {
+        it('should show all the rows', function () {
+          const tableBodyRows = queryAllByRole('row')
+            .filter((row) => (
+              within(row).queryAllByRole('columnheader').length === 0
+            ));
+          strictEqual(tableBodyRows.length, testFacultySchedules.length);
+        });
+      });
+      context('when the filter is changed', function () {
+        it('should show the rows with first names matching the filters', function () {
+          const testFirstName = testFacultySchedules[0].firstName;
+          const filter = getByLabelText(firstNameLabel);
+          fireEvent.change(
+            filter,
+            {
+              target: {
+                value: testFirstName,
+              },
+            }
+          );
+          const expectedNumRows = testFacultySchedules
+            .filter((faculty) => faculty.firstName === testFirstName).length;
+          const tableBodyRows = queryAllByRole('row')
+            .filter((row) => (
+              within(row).queryAllByRole('columnheader').length === 0
+            ));
+          const actualNumRows = tableBodyRows.map((row) => (
+            row.textContent
+          )).length;
+          strictEqual(expectedNumRows, actualNumRows);
+        });
+      });
+    });
+    describe('Category Filter', function () {
+      context('when no filters have been changed', function () {
+        it('should show all the rows', function () {
+          const tableBodyRows = queryAllByRole('row')
+            .filter((row) => (
+              within(row).queryAllByRole('columnheader').length === 0
+            ));
+          strictEqual(tableBodyRows.length, testFacultySchedules.length);
+        });
+        it('should show all of the categories and "All" as options', function () {
+          const filter = getByLabelText('Change to filter the faculty list by category') as HTMLSelectElement;
+          const options = within(filter)
+            .getAllByRole('option') as HTMLOptionElement[];
+          const actualOptions = options
+            .map(({ textContent }) => textContent);
+          const expectedOptions = Object.values(FACULTY_TYPE).map((type) => (
+            facultyTypeEnumToTitleCase(type)
+          ));
+          // The default value of the area dropdown is 'All'
+          expectedOptions.unshift('All');
+          deepStrictEqual(actualOptions, expectedOptions);
+        });
+      });
+      context('when the filter is changed', function () {
+        it('should show the rows with the category matching the filter', function () {
+          const testCategory = testFacultySchedules[0].category;
+          const filter = getByLabelText(categoryLabel);
+          fireEvent.change(
+            filter,
+            {
+              target: {
+                value: testCategory,
+              },
+            }
+          );
+          const expectedNumRows = testFacultySchedules
+            .filter((faculty) => faculty.category === testCategory).length;
+          const tableBodyRows = queryAllByRole('row')
+            .filter((row) => (
+              within(row).queryAllByRole('columnheader').length === 0
+            ));
+          const actualNumRows = tableBodyRows.map((row) => (
+            row.textContent
+          )).length;
+          strictEqual(expectedNumRows, actualNumRows);
+        });
+      });
+    });
+    describe('Fall Absence Filter', function () {
+      context('when no filters have been changed', function () {
+        it('should default to the value "All"', function () {
+          const rows = getAllByRole('row');
+          const utils = within(rows[2]);
+          const filter = utils
+            .queryByLabelText(fallAbsenceLabel) as HTMLSelectElement;
+          strictEqual(filter.value, 'All');
+        });
+        it('should show all of the absence types and "All" as options', function () {
+          const rows = getAllByRole('row');
+          const utils = within(rows[2]);
+          const filter = utils
+            .queryByLabelText(fallAbsenceLabel) as HTMLSelectElement;
+          const options = within(filter)
+            .getAllByRole('option') as HTMLOptionElement[];
+          const actualOptions = options
+            .map(({ textContent }) => textContent);
+          const expectedOptions = Object.values(ABSENCE_TYPE).map((type) => (
+            absenceEnumToTitleCase(type)
+          ));
+          // The default value of the area dropdown is 'All'
+          expectedOptions.unshift('All');
+          deepStrictEqual(actualOptions, expectedOptions);
+        });
+        it('should show all the rows', function () {
+          const tableBodyRows = queryAllByRole('row')
+            .filter((row) => (
+              within(row).queryAllByRole('columnheader').length === 0
+            ));
+          strictEqual(tableBodyRows.length, testFacultySchedules.length);
+        });
+      });
+      context('when the filter is changed', function () {
+        it('should show the rows matching the fall absence value selected', function () {
+          const testAbsence = testFacultySchedules[0].fall.absence.type;
+          const filter = getByLabelText(fallAbsenceLabel);
+          fireEvent.change(
+            filter,
+            {
+              target: {
+                value: testAbsence,
+              },
+            }
+          );
+          const expectedNumRows = testFacultySchedules
+            .filter((faculty) => faculty.fall.absence.type
+            === testAbsence).length;
+          const tableBodyRows = queryAllByRole('row')
+            .filter((row) => (
+              within(row).queryAllByRole('columnheader').length === 0
+            ));
+          const actualNumRows = tableBodyRows.map((row) => (
+            row.textContent
+          )).length;
+          strictEqual(expectedNumRows, actualNumRows);
+        });
+      });
+    });
+    describe('Spring Absence Filter', function () {
+      context('when no filters have been changed', function () {
+        it('should default to the value "All"', function () {
+          const rows = getAllByRole('row');
+          const utils = within(rows[2]);
+          const filter = utils
+            .queryByLabelText(springAbsenceLabel) as HTMLSelectElement;
+          strictEqual(filter.value, 'All');
+        });
+        it('should show all of the absence types and "All" as options', function () {
+          const rows = getAllByRole('row');
+          const utils = within(rows[2]);
+          const filter = utils
+            .queryByLabelText(springAbsenceLabel) as HTMLSelectElement;
+          const options = within(filter)
+            .getAllByRole('option') as HTMLOptionElement[];
+          const actualOptions = options
+            .map(({ textContent }) => textContent);
+          const expectedOptions = Object.values(ABSENCE_TYPE).map((type) => (
+            absenceEnumToTitleCase(type)
+          ));
+          // The default value of the area dropdown is 'All'
+          expectedOptions.unshift('All');
+          deepStrictEqual(actualOptions, expectedOptions);
+        });
+        it('should show all the rows', function () {
+          const tableBodyRows = queryAllByRole('row')
+            .filter((row) => (
+              within(row).queryAllByRole('columnheader').length === 0
+            ));
+          strictEqual(tableBodyRows.length, testFacultySchedules.length);
+        });
+      });
+      context('when the filter is changed', function () {
+        it('should show the rows matching the spring absence value selected', function () {
+          const testAbsence = testFacultySchedules[0].spring.absence.type;
+          const filter = getByLabelText(springAbsenceLabel);
+          fireEvent.change(
+            filter,
+            {
+              target: {
+                value: testAbsence,
+              },
+            }
+          );
+          const expectedNumRows = testFacultySchedules
+            .filter((faculty) => faculty.spring.absence.type
+          === testAbsence).length;
+          const tableBodyRows = queryAllByRole('row')
+            .filter((row) => (
+              within(row).queryAllByRole('columnheader').length === 0
+            ));
+          const actualNumRows = tableBodyRows.map((row) => (
+            row.textContent
+          )).length;
+          strictEqual(expectedNumRows, actualNumRows);
+        });
+      });
     });
   });
 });
